+++ src/dymaptic.GeoBlazor.Core/Components/Layers/FeatureLayer.cs
@@ -83,8 +83,8 @@ public class FeatureLayer : Layer, IFeatureReductionLayer, IPopupTemplateLayer, 
     ///     The <see cref="PopupTemplate" /> for the layer.
     /// </param>
     public FeatureLayer(string? url = null, PortalItem? portalItem = null, IReadOnlyList<Graphic>? source = null,
-        string[]? outFields = null, string? definitionExpression = null, double? minScale = null,
-        double? maxScale = null, string? objectIdField = null, GeometryType? geometryType = null, string? title = null,
+        string[]? outFields = null, string? definitionExpression = null, int? minScale = null,
+        int? maxScale = null, string? objectIdField = null, GeometryType? geometryType = null, string? title = null,
         double? opacity = null, bool? visible = null, ListMode? listMode = null, PopupTemplate? popupTemplate = null)
     {
         if (url is null && portalItem is null && source is null)
@@ -158,14 +158,14 @@ public class FeatureLayer : Layer, IFeatureReductionLayer, IPopupTemplateLayer, 
     /// </summary>
     [Parameter]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public double? MinScale { get; set; }
+    public int? MinScale { get; set; }
 
     /// <summary>
     ///     The maximum scale (most zoomed in) at which the layer is visible in the view.
     /// </summary>
     [Parameter]
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public double? MaxScale { get; set; }
+    public int? MaxScale { get; set; }
 
     /// <summary>
     ///     The name of an oidfield containing a unique value or identifier for each feature in the layer.
